# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LNT Paris Stock Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Session {
  id: number;
  date: string;
  heure: string;
  localisation?: string | null;
  typePaiement: string;
  createdAt: string;
}

export interface CreateSession {
  date: string;
  heure: string;
  localisation?: string | null;
  typePaiement: string;
}

export interface Collection {
  id: number;
  nom: string;
  description?: string | null;
  createdAt: string;
}

export interface Produit {
  id: number;
  collectionId: number;
  couleur: string;
  quantite: number;
  prixCentimes: number;
  createdAt: string;
}

export interface CollectionWithProduits {
  id: number;
  nom: string;
  description?: string | null;
  createdAt: string;
  produits: Produit[];
}

export interface CreateCollection {
  nom: string;
  description?: string | null;
}

export interface CreateProduit {
  collectionId: number;
  couleur: string;
  quantite: number;
  prixCentimes: number;
}

export interface UpdateProduit {
  quantite?: number;
  couleur?: string;
  prixCentimes?: number;
}

export interface CreateVente {
  produitId: number;
  quantiteVendue: number;
  typePaiement: string;
}

export interface Vente {
  id: number;
  produitId: number;
  quantiteVendue: number;
  typePaiement: string;
  montantCentimes: number;
  createdAt: string;
}

export interface SuccessMessage {
  message: string;
}