# lib/api-zod/src/generated/types/createProduit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LNT Paris Stock Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProduit {
  collectionId: number;
  couleur: string;
  quantite: number;
  prixCentimes: number;
}