# lib/api-zod/src/generated/types/vente.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LNT Paris Stock Management API
 * OpenAPI spec version: 0.1.0
 */

export interface Vente {
  id: number;
  produitId: number;
  quantiteVendue: number;
  typePaiement: string;
  montantCentimes: number;
  createdAt: string;
}